Reorganize code and add command support
Authored-by: Daniel Keeney <dkeeney@pivotal.io>
