+++ jobs/print-component-version/templates/run.erb
@@ -4,4 +4,4 @@ set -euo pipefail
 
 cd /var/vcap/jobs/print-component-version/
 
-./packages/osl-collector/osl-collector -folder ${PWD}/packages
+./packages/osl-collector/osl-collector -folder ${PWD}/packages collect
