+++ src/osl-collector/commands/osm.go
@@ -1,0 +1,7 @@
+package commands
+
+import "fmt"
+
+func Osm() {
+	fmt.Println("Not yet implemented")
+}
