+++ src/osl-collector/core/oslStructs/structs.go
