+++ src/osl-collector/core/oslStructs/structs_test.go
