+++ src/osl-collector/core/testNested/testDir2/boshcycle_osm.json
